% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm.R
\name{coef.asm}
\alias{coef.asm}
\title{Coefficients of an \code{asm} regression model}
\usage{
\method{coef}{asm}(object, ...)
}
\arguments{
\item{object}{asm object}

\item{...}{additional arguments to ensure compatibility with the generic function coef()}
}
\value{
vector of coefficients of the \code{asm} regression model
}
\description{
Outputs the coefficients of a fitted \code{asm} regression model
}
\examples{
model = asm(mpg ~ cyl + hp + disp, data=mtcars)
coef(model)

}
