% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{fd_est_asm}
\alias{fd_est_asm}
\title{Estimate the Active Subspace of a Function using Finite Difference Gradients}
\usage{
fd_est_asm(f, r, m, M = NULL, scale = FALSE)
}
\arguments{
\item{f}{The function to eval}

\item{r}{The max dim of the active subspace}

\item{m}{The dimension of the underlying/embedding space.}

\item{scale}{Scale all gradients to have norm 1?}

\item{thresh}{Bound on angle between subspaces.}
}
\value{
A list with sub, the active subspace, sv, the singular values (all m of them), fs, which gives function values, gs, function grads, and X, which gives sampled locations.
}
\description{
Looks between [-1, 1]
}
