% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm.R
\name{print.asm}
\alias{print.asm}
\title{Short description of a fitted \code{asm} regression model}
\usage{
\method{print}{asm}(x, ...)
}
\arguments{
\item{x}{asm object}

\item{...}{additional arguments to ensure compatibility with the generic function print()}
}
\value{
No return value, called for its side effect
}
\description{
Outputs estimated coefficients and standard errors
}
\examples{
model = asm(mpg ~ cyl + hp + disp, data=mtcars)
print(model)
}
