% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asm.R
\name{residuals.asm}
\alias{residuals.asm}
\title{Residuals from an \code{asm} regression model}
\usage{
\method{residuals}{asm}(object, ...)
}
\arguments{
\item{object}{asm object}

\item{...}{additional arguments to ensure compatibility with the generic function residuals()}
}
\value{
vector of residuals from the \code{asm} regression model
}
\description{
Outputs the residuals (on the training data) from a fitted \code{asm} regression
             model
}
\examples{
model = asm(mpg ~ cyl + hp + disp, data=mtcars)
residuals(model)

}
