% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{sel_asm}
\alias{sel_asm}
\title{Automatically Select an Active Subspace}
\usage{
sel_asm(sv, method = "both", b = exp(1), thresh = 1, alpha = 0.8)
}
\arguments{
\item{sv}{A vector of singular values.}

\item{method}{One of 'both', 'log_gap', or 'perc_var'. See Details.}

\item{b}{The log base for the 'log_gap' method.}

\item{thresh}{The threshold for the 'log_gap' method.}

\item{alpha}{The proportion of the total that must be exceeded for the 'perc_var' method.}
}
\value{
The index of the smallest singular value to be retained, or NULL if none pass the tests.
}
\description{
Select an active subspace via two methods (or both). Method 'log_gap' Examines for an absolute difference in the base b logarithm of the singular values of greater than thresh. Method 'perc_var' picks singular values until alpha % of the total has been accumulated. Method 'both' selects singular values passing both 'log_gap' and 'perc_var' tests.
}
