% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{subspace_dist}
\alias{subspace_dist}
\title{Get the distance between subspaces defined as the ranges of A and B}
\usage{
subspace_dist(A, B)
}
\arguments{
\item{A}{A matrix}

\item{B}{Another matrix with the same number of rows as A.}
}
\value{
A nonnegative scalar given the norm of the difference between the projectors onto the respective subspaces.
}
\description{
Get the distance between subspaces defined as the ranges of A and B
}
